requirements.txt
langchain>=0.1.0
langchain-community>=0.0.10
langchain-core>=0.1.0
langchain-ollama>=0.1.0
langchain-huggingface>=0.0.5
langchain-chroma>=0.0.5
sentence-transformers==2.2.2
chromadb==0.4.18
gradio==4.44.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
plotly==5.18.0
pandas==2.1.4
pytest==7.4.3
pydantic>=2.5.0
transformers>=4.30.0
huggingface-hub==0.17.3